StudentDashbord is a disaster zone of permission checks that Android Studio keeps screaming about - and even though they're in it yells at us because the permissions don't occur inside the main function. Whatever the case GPS is now a possibility but we need some sort of library on the Raspberry Pi that can do the same thing and get accurate coordinates so we can take distances.
Administrator Dashboard now correctly links to different activities but they're all empty so who cares until it actually works.

Network communication on the way but currently non-fuctional.

Fixed a problem where the student dashboard would infinitely switch back and forth between printing the "FAIL" and "SUCCEED" texts. It's not like it wastes power or time or anything BECAUSE WHO DOES THAT!?