This marks the point at which Networking has begun. Login & Register both have networking capabilities, we now need to determine the IP address of the Raspberry Pi and an open Port through which we can communicate.
As of right now it is impossible to test as to whether or not the App will function because while Networking is active on the App's side, the Database on the Raspberry Pi is not set up and the IP that it uses is not known.

If you want to test the functions for yourself then you must clone an older commit (Preferably the one dirctly before this one).

RaspberryPiCommunication.java handles all of the communications from the Raspberry Pi to and from.