Strange AndroidStudio problems
For whatever reason Android Studio does not recognize this as a java file even  though it clearly is. It didn't give any errors when it should of, but those are fixed now. Obviously we're throwing a crapton of IOExceptions that  need to be handled. Documentation still needs to be done.