PART TWO OF THE FINALE BETWEEN ME AND NETWORKING.
A lot of the manual I wrote in RaspberryPiCommunication.java will need to be replaced in the very near future, but I just reconciled thousands of years of theology, physics, psychology, and neurology. In other words, praying, punching the monitor, begging my computer, and crying tears of joy when this god awful monstrosity came to life and serviced my requests without hesitations.

It's like midnight right now. On a Friday.