IT WORKS, THE CIRCLE WORKS, OH MY GOD. OH MY GOD.
We no longer sit for 50+ seconds waiting for a server. We now wait 3 seconds to find a server. If we can not access it/whatever happens, then we kill the connection and return. On the UI side of things, we now have a progress circle that lets the user know that the app has not yet crashed.